+++ lib/jets/commands/templates/skeleton/spec/spec_helper.rb.tt
@@ -10,7 +10,7 @@ require "jets"
 abort("The Jets environment is running in production mode!") if Jets.env == "production"
 Jets.boot
 
-require "jets/spec/helpers"
+require "jets/spec_helpers"
 
 <% if @webpacker -%>
 require 'capybara/rspec'
