+++ lib/jets/spec_helpers/response.rb
@@ -1,0 +1,5 @@
+module Jets
+  module SpecHelpers
+    Response = Struct.new(:status, :body)
+  end
+end
